requirements.txt
opencv-python
ultralytics
google-generativeai
python-dotenv
Pillow
numpy
google-adk
fastapi
uvicorn
